+++ CHANGELOG.md
@@ -32,6 +32,8 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Simplify and speed up case general report code
 - Speed up case retrieval in case_matching_causatives
 - Upgrade pymongo to version 4
+- When updating disease terms, check that all terms are consistent with a DiseaseTerm model before dropping the old collection
+- Better separation between modules loading HPO terms and diseases
 
 
 ## [4.71]